Patchnotes v1.01
Changed some comments in WhileLoopFun.java

Added RunningAverage.java class, it calculates the average of a set of numbers that the user inputs
Average() takes user input and adds it to a total sum while incrementing a counter by 1. After the user ends the program, it calculates the average of the numbers entered by dividing the sum of all the numbers by the count of how many times the user has entered numbers, and prints a statement according to its results.